Cleanup the threading, its a method in openfiremanager now to spawn a threaded broadcast.
Also added date to the broadcast.
